API helper and refactors
