+++ frontend/src/models/connections.ts
@@ -1,10 +1,11 @@
 import { createModel } from '@rematch/core'
-import { API_URL, DEVELOPER_KEY } from '../shared/constants'
+import { DEVELOPER_KEY } from '../shared/constants'
 import { newConnection, setConnection } from '../helpers/connectionHelper'
 import { r3, getToken } from '../services/remote.it'
 import { selectById } from '../models/devices'
 import { RootModel } from './rootModel'
 import axios from 'axios'
+import { getApiURL } from '../helpers/apiHelper'
 
 type ConnectionsState = {
   all: IConnection[]
@@ -73,7 +74,7 @@ export default createModel<RootModel>()({
       setConnection(proxyConnection)
 
       try {
-        let result = await axios.post(`${API_URL}/device/connect`, data, await dispatch.connections.headerOptions())
+        let result = await axios.post(`${getApiURL()}/device/connect`, data, await dispatch.connections.headerOptions())
         const response = r3.processData(result)
         const proxyResult: ProxyConnectionResult = response.connection || {}
         console.log('PROXY CONNECTED', proxyResult)
@@ -103,7 +104,7 @@ export default createModel<RootModel>()({
       const data = { deviceaddress: connection.id, connectionid: connection.publicId }
       try {
         let result = await axios.post(
-          `${API_URL}/device/connect/stop`,
+          `${getApiURL()}/device/connect/stop`,
           data,
           await dispatch.connections.headerOptions()
         )
