+++ frontend/src/pages/TestPage.tsx
@@ -5,7 +5,7 @@ import { useSelector, useDispatch } from 'react-redux'
 import { ListItemSetting } from '../components/ListItemSetting'
 import { Container } from '../components/Container'
 import { TestUI } from '../components/TestUI'
-import { getApiURL } from '../services/graphQL'
+import { getApiURL } from '../helpers/apiHelper'
 import { Title } from '../components/Title'
 import { Quote } from '../components/Quote'
 import { emit } from '../services/Controller'
