+++ frontend/src/services/graphQL.ts
@@ -1,21 +1,7 @@
 import axios, { AxiosError, AxiosResponse } from 'axios'
 import { getToken } from '../services/remote.it'
 import { store } from '../store'
-import { GRAPHQL_API, GRAPHQL_BETA_API } from '../shared/constants'
-import { version } from '../../package.json'
-
-export function getApiURL(): string {
-  const { apiGraphqlURL, switchApi } = store.getState().backend.preferences
-  const { overrides } = store.getState().backend.environment
-
-  let beta = version.includes('alpha')
-
-  return apiGraphqlURL && switchApi
-    ? apiGraphqlURL
-    : beta
-    ? overrides?.betaApiURL || GRAPHQL_BETA_API
-    : overrides?.apiURL || GRAPHQL_API
-}
+import { getApiURL } from '../helpers/apiHelper'
 
 export async function graphQLBasicRequest(query: String, variables: ILookup<any> = {}) {
   try {
